refactor: move video components to wrapper component
